requirements.txt
pytest-sanic==1.6.1
sanic==20.3.0
alembic==1.4.2